project/requirements.txt
strands-agents
strands-agents-tools
boto3>=1.40.8
botocore>=1.40.8
bedrock-agentcore>=0.1.2
bedrock-agentcore-starter-toolkit>=0.1.5
aws-opentelemetry-distro~=0.10.1
ddgs
pyyaml
python-dotenv
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
jinja2>=3.1.0